Added a centralized ExerciseService and AuthService
Exercise Service - This service enables users to start a specific service based on the needs of the user which is dynamic in itself.
It storethe private exercise data array , so that any component can access this data via a getter method as , no one should be able to alter the orignal database.
Special Components like Subject and Subscription are also used to manage data array passing between the components which makes it more flexible 

AuthService - This service currently behaves a temporary Authorization service for a user registering and logging in.
Note : TrainingComponent is loaded only when the user is logged in.
Otherwise it doesnt allow users to go that link .
Thanks to the CanActivateGuard by Angular

Further Improvements => Connecting and hosting this App to actual Realtime database . thus eliminating the temporary Authorization Service