Hosted the App on cloud , Added Strong User Authentication &  Data Storage on cloud
As decided previously , we had cloud as well as replacement of temporary authorization code into a actual user Authentication code in our pipeline .
In this update , entire App is hosted on firebase with functionalities , User Authentication , Security as well data is stored in the database.
Optimization and restructuring of entire code has been done.
The is a stable release